+++ admin/js.js
@@ -427,10 +427,15 @@ function efml_sync_save_config() {
   });
   jQuery('.eml-sync-config input[type="checkbox"]').each(function(){
     if( jQuery(this).is(':checked') ) {
-      if( ! fields[jQuery( this ).attr( 'name' ).replace('[', '').replace(']', '')] ) {
-        fields[jQuery( this ).attr( 'name' ).replace('[', '').replace(']', '')] = {};
+      if( jQuery( this ).attr( 'name' ).indexOf('[') >= 0 ) {
+        if (!fields[jQuery( this ).attr( 'name' ).replace( '[', '' ).replace( ']', '' )]) {
+          fields[jQuery( this ).attr( 'name' ).replace( '[', '' ).replace( ']', '' )] = {};
+        }
+        fields[jQuery( this ).attr( 'name' ).replace( '[', '' ).replace( ']', '' )][jQuery( this ).val()] = 1;
       }
-      fields[jQuery( this ).attr( 'name' ).replace('[', '').replace(']', '')][jQuery( this ).val()] = 1;
+      else {
+        fields[jQuery( this ).attr( 'name' )] = 1;
+      }
     }
   });
   fields['term_id'] = jQuery('#term_id').val();
@@ -448,23 +453,7 @@ function efml_sync_save_config() {
       efml_ajax_error_dialog( errorThrown )
     },
     success: function (response) {
-      let dialog_config = {
-        detail: {
-          className: 'eml',
-          title: efmlJsVars.title_sync_config_saved,
-          texts: [
-            '<p>' + efmlJsVars.text_sync_config_saved + '</p>'
-          ],
-          buttons: [
-            {
-              'action': 'location.reload();',
-              'variant': 'primary',
-              'text': efmlJsVars.lbl_ok
-            },
-          ]
-        }
-      }
-      efml_create_dialog( dialog_config );
+      efml_create_dialog( response );
     }
   });
 }